requirements.txt
# Core Django
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Firebase
firebase-admin==6.3.0

# Background tasks
celery==5.3.4
redis==5.0.1

# Database
psycopg2-binary==2.9.9

# Authentication & Security
djangorestframework-simplejwt==5.3.1
python-decouple==3.8
cryptography==41.0.7

# Utils
Pillow==10.2.0
python-dateutil==2.8.2

# API Documentation
drf-yasg==1.21.7

# Internationalization
django-rosetta==0.10.0

# Development
django-debug-toolbar==4.2.0

# Testing
pytest==7.4.4
pytest-django==4.7.0
pytest-cov==4.1.0

# Code quality
black==23.12.1
flake8==7.0.0
isort==5.13.2